Extract expression formatting out of KotlinInputAstVisitor into ExpressionFormatter
Purely mechanical change that extracts all `visit*` functions related to basic (non control flow) expressions into corresponding `format*` analogs
