+++ core/src/main/kotlin/org/jetbrains/ktfmt/format/visitor/FileFormatter.kt
@@ -5,6 +5,8 @@ import org.jetbrains.kotlin.com.intellij.psi.PsiComment
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.kotlin.psi.KtFile
+import org.jetbrains.kotlin.psi.KtImportDirective
+import org.jetbrains.kotlin.psi.KtPackageDirective
 import org.jetbrains.kotlin.psi.KtProperty
 import org.jetbrains.kotlin.psi.KtScript
 import org.jetbrains.kotlin.psi.KtScriptInitializer
@@ -69,5 +71,57 @@ interface FileFormatter : KotlinAstFormatter {
       formatStatement(statement)
       markForPartialFormat()
     }
+  }
+
+  override fun formatPackageDirective(directive: KtPackageDirective) {
+    builder.sync(directive)
+    if (directive.packageKeyword == null) {
+      return
+    }
+    builder.token("package")
+    builder.space()
+    var first = true
+    for (packageName in directive.packageNames) {
+      if (first) {
+        first = false
+      } else {
+        builder.token(".")
+      }
+      builder.token(packageName.getIdentifier()?.text ?: packageName.getReferencedName())
+    }
+
+    builder.guessToken(";")
+    builder.forcedBreak()
+  }
+
+  /** Example `import com.foo.A` */
+  override fun formatImportDirective(directive: KtImportDirective) {
+    builder.sync(directive)
+    builder.token("import")
+    builder.space()
+
+    val importedReference = directive.importedReference
+    if (importedReference != null) {
+      inImport = true
+      format(importedReference)
+      inImport = false
+    }
+    if (directive.isAllUnder) {
+      builder.token(".")
+      builder.token("*")
+    }
+
+    // Possible alias.
+    val alias = directive.alias?.nameIdentifier
+    if (alias != null) {
+      builder.space()
+      builder.token("as")
+      builder.space()
+      builder.token(alias.text ?: fail())
+    }
+
+    // Force a newline afterwards.
+    builder.guessToken(";")
+    builder.forcedBreak()
   }
 }
